+++ workspaces/scorecard/plugins/scorecard-backend-module-jira/src/metricProviders/JiraOpenIssuesProvider.ts
@@ -16,7 +16,7 @@
 
 import type { Config } from '@backstage/config';
 import type { Entity } from '@backstage/catalog-model';
-import { THRESHOLDS_CONFIG_PATH } from '../constants';
+import { JIRA_CONFIG_PATH, THRESHOLDS_CONFIG_PATH } from '../constants';
 import {
   DEFAULT_NUMBER_THRESHOLDS,
   Metric,
@@ -28,13 +28,27 @@ import {
 } from '@red-hat-developer-hub/backstage-plugin-scorecard-node';
 import { JiraClient } from '../clients/base';
 import { JiraClientFactory } from '../clients/JiraClientFactory';
+import { ScorecardJiraAnnotations } from '../annotations';
+import { AuthService, DiscoveryService } from '@backstage/backend-plugin-api';
+import {
+  ConnectionStrategy,
+  DirectConnectionStrategy,
+  ProxyConnectionStrategy,
+} from '../strategies/ConnectionStrategy';
+import { Product } from '../clients/types';
+
+const { PROJECT_KEY } = ScorecardJiraAnnotations;
 
 export class JiraOpenIssuesProvider implements MetricProvider<'number'> {
   private readonly thresholds: ThresholdConfig;
   private readonly jiraClient: JiraClient;
 
-  private constructor(config: Config, thresholds?: ThresholdConfig) {
-    this.jiraClient = JiraClientFactory.create(config);
+  private constructor(
+    config: Config,
+    connectionStrategy: ConnectionStrategy,
+    thresholds?: ThresholdConfig,
+  ) {
+    this.jiraClient = JiraClientFactory.create(config, connectionStrategy);
     this.thresholds = thresholds ?? DEFAULT_NUMBER_THRESHOLDS;
   }
 
@@ -43,7 +57,7 @@ export class JiraOpenIssuesProvider implements MetricProvider<'number'> {
   }
 
   getProviderId() {
-    return 'jira.open-issues';
+    return 'jira.open_issues';
   }
 
   getMetric(): Metric<'number'> {
@@ -62,16 +76,45 @@ export class JiraOpenIssuesProvider implements MetricProvider<'number'> {
   }
 
   supportsEntity(entity: Entity): boolean {
-    return entity.metadata.annotations?.['jira/project-key'] !== undefined;
+    return entity.metadata.annotations?.[PROJECT_KEY] !== undefined;
   }
 
-  static fromConfig(config: Config): JiraOpenIssuesProvider {
+  static fromConfig(
+    config: Config,
+    options: {
+      auth: AuthService;
+      discovery: DiscoveryService;
+    },
+  ): JiraOpenIssuesProvider {
     const configuredThresholds = config.getOptional(THRESHOLDS_CONFIG_PATH);
     if (configuredThresholds !== undefined) {
       validateThresholds(configuredThresholds, 'number');
     }
 
-    return new JiraOpenIssuesProvider(config, configuredThresholds);
+    let connectionStrategy: ConnectionStrategy;
+
+    const jiraConfig = config.getConfig(JIRA_CONFIG_PATH);
+    const proxyPath = jiraConfig.getOptionalString('proxyPath');
+
+    if (proxyPath) {
+      connectionStrategy = new ProxyConnectionStrategy(
+        proxyPath,
+        options.auth,
+        options.discovery,
+      );
+    } else {
+      connectionStrategy = new DirectConnectionStrategy(
+        jiraConfig.getString('baseUrl'),
+        jiraConfig.getString('token'),
+        jiraConfig.getString('product') as Product,
+      );
+    }
+
+    return new JiraOpenIssuesProvider(
+      config,
+      connectionStrategy,
+      configuredThresholds,
+    );
   }
 
   async calculateMetric(entity: Entity): Promise<number> {